+++ apps/web/modules/api/v2/me/route.ts
@@ -1,7 +1,7 @@
 import { authenticatedApiClient } from "@/modules/api/v2/auth/authenticated-api-client";
 import { responses } from "@/modules/api/v2/lib/response";
 import { handleApiError } from "@/modules/api/v2/lib/utils";
-import { hasOrganizationAccess } from "@/modules/api/v2/organizations/[organizationId]/lib/utils";
+import { hasOrganizationAccess } from "@/modules/organization/settings/api-keys/lib/utils";
 import { NextRequest } from "next/server";
 import { OrganizationAccessType } from "@formbricks/types/api-key";
