+++ apps/web/modules/api/v2/organizations/[organizationId]/lib/utils.ts
@@ -1,22 +1,7 @@
+import { hasOrganizationAccess } from "@/modules/organization/settings/api-keys/lib/utils";
 import { logger } from "@formbricks/logger";
 import { OrganizationAccessType } from "@formbricks/types/api-key";
 import { TAuthenticationApiKey } from "@formbricks/types/auth";
-
-export const hasOrganizationAccess = (
-  authentication: TAuthenticationApiKey,
-  accessType: OrganizationAccessType
-): boolean => {
-  const organizationAccess = authentication.organizationAccess?.accessControl;
-
-  switch (accessType) {
-    case OrganizationAccessType.Read:
-      return organizationAccess?.read === true || organizationAccess?.write === true;
-    case OrganizationAccessType.Write:
-      return organizationAccess?.write === true;
-    default:
-      return false;
-  }
-};
 
 export const hasOrganizationIdAndAccess = (
   paramOrganizationId: string,
