+++ apps/web/modules/organization/settings/api-keys/lib/utils.ts
@@ -1,5 +1,6 @@
 import { TFnType } from "@tolgee/react";
-import { TAPIKeyEnvironmentPermission } from "@formbricks/types/auth";
+import { OrganizationAccessType } from "@formbricks/types/api-key";
+import { TAPIKeyEnvironmentPermission, TAuthenticationApiKey } from "@formbricks/types/auth";
 
 // Permission level required for different HTTP methods
 const methodPermissionMap = {
@@ -48,5 +49,21 @@ export const getOrganizationAccessKeyDisplayName = (key: string, t: TFnType) => 
       return t("environments.project.api_keys.access_control");
     default:
       return key;
+  }
+};
+
+export const hasOrganizationAccess = (
+  authentication: TAuthenticationApiKey,
+  accessType: OrganizationAccessType
+): boolean => {
+  const organizationAccess = authentication.organizationAccess?.accessControl;
+
+  switch (accessType) {
+    case OrganizationAccessType.Read:
+      return organizationAccess?.read === true || organizationAccess?.write === true;
+    case OrganizationAccessType.Write:
+      return organizationAccess?.write === true;
+    default:
+      return false;
   }
 };
